@@ -3,14 +3,14 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:intl/intl.dart';
-import 'package:weechat/pages/channel/urlify.dart';
+import 'package:weechat/widgets/channel/urlify.dart';
 import 'package:weechat/relay/buffer.dart';
 import 'package:weechat/relay/colors.dart';
 
 class LineItem extends StatelessWidget {
   final LineData line;
 
-  LineItem({required this.line});
+  const LineItem({required this.line, Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -22,27 +22,25 @@ class LineItem extends StatelessWidget {
     final isSystem = line.prefix.isEmpty ||
         ['<--', '-->', '--', '===', '=!='].any((e) => line.prefix.endsWith(e));
     final alpha = isSystem ? 100 : 255;
-    final defaultColor = tt.bodyText2?.color ?? th.colorScheme.onSurface;
+    final defaultColor = tt.bodyMedium?.color ?? th.colorScheme.onSurface;
 
     final isAction = [' *'].any((e) => line.prefix.endsWith(e));
-    var bodyStyle = tt.bodyText2;
+    var bodyStyle = tt.bodyMedium;
     if (isAction) bodyStyle = bodyStyle?.copyWith(fontStyle: FontStyle.italic);
 
-    //print('<${line.prefix}> ${line.message} (${line.message.codeUnits.map((e) => e.toRadixString(16)).toList()})');
-
-    final prefixRT = parseColors(line.prefix, defaultColor, alpha: alpha).text;
+    final prefixRT = parseColors(line.prefix, defaultColor, alpha: alpha);
     final messageRT =
         parseColors(line.message, defaultColor, alpha: alpha).text as TextSpan;
 
     final dateRT = Container(
-      padding: EdgeInsets.symmetric(vertical: 1, horizontal: 3),
-      margin: EdgeInsets.only(right: 5),
+      padding: const EdgeInsets.symmetric(vertical: 1, horizontal: 3),
+      margin: const EdgeInsets.only(right: 5),
       color: line.highlight ? Colors.redAccent : null,
       child: RichText(
         text: TextSpan(
-          text: '$df',
-          style: tt.bodyText2?.copyWith(
-            fontFeatures: [FontFeature.tabularFigures()],
+          text: df,
+          style: tt.bodyMedium?.copyWith(
+            fontFeatures: [const FontFeature.tabularFigures()],
             color: line.highlight ? Colors.white : Colors.grey.withAlpha(100),
           ),
         ),
@@ -53,10 +51,10 @@ class LineItem extends StatelessWidget {
       text: TextSpan(
         style: bodyStyle,
         children: [
-          if (!(isSystem || isAction)) TextSpan(text: '<', style: tt.bodyText2),
-          if (!isAction) prefixRT,
+          if (!(isSystem || isAction)) TextSpan(text: '<', style: tt.bodyMedium),
+          if (!isAction) prefixRT.text,
           if (!isAction && line.prefix.isNotEmpty)
-            TextSpan(text: isSystem ? ' ' : '> ', style: tt.bodyText2),
+            TextSpan(text: isSystem ? ' ' : '> ', style: tt.bodyMedium),
           urlify(messageRT, onNotification: (msg) {
             ScaffoldMessenger.of(context).showSnackBar(SnackBar(
               content: Text(msg),
@@ -67,7 +65,7 @@ class LineItem extends StatelessWidget {
     );
 
     return Container(
-      padding: EdgeInsets.only(top: 3),
+      padding: const EdgeInsets.only(top: 3),
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [dateRT, Expanded(child: bodyRT)],
