@@ -5,6 +5,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:weechat/relay/colors/color_code_parser.dart';
+import 'package:weechat/relay/colors/color_codes.dart';
 import 'package:weechat/relay/colors/rich_text_parser.dart';
 
 RichText parseColors(
@@ -12,7 +13,12 @@ RichText parseColors(
   Color defaultColor, {
   TextStyle? textStyle,
   int? alpha,
+  ColorCodes? colorCodes,
+  ColorOptions? colorOptions,
 }) {
+  colorCodes ??= defaultColorCodes;
+  colorOptions ??= defaultColorOptions;
+
   final it = raw.runes.iterator;
 
   final p = RichTextParser(defaultFgColor: defaultColor, defaultAlpha: alpha);
@@ -42,12 +48,14 @@ RichText parseColors(
     else if (it.current == 0x19) {
       p.finalizeCurrentSpan();
       ColorCodeParser ccp = ColorCodeParser(defaultFgColor: defaultColor);
-      if (ccp.parse(it)) {
+      if (ccp.parse(it, colorCodes: colorCodes, colorOptions: colorOptions)) {
         if (ccp.fgColor != null) p.fgColor = ccp.fgColor;
         if (ccp.bgColor != null) p.bgColor = ccp.bgColor;
         if (ccp.attributes != null) p.attributes.set(ccp.attributes!);
       }
-    } else if (it.current > 0) p.addText(it.currentAsString);
+    } else if (it.current > 0) {
+      p.addText(it.currentAsString);
+    }
   }
 
   p.finalizeCurrentSpan();
